[feat] Add climber control (see desc)
LStick and RStick control the respective climbers (up/down, slower/faster) based on their respective Y values at the time that they are clicked. Moving the stick around while the button is held doesn't change the speed due to limitations of the lambda expression. I suggest allowing the operator to test as-is, and if it's cumbersome, we can look into restructuring the Climber commands.

Also, commented out the lines setting up the right climber motor as a follower. I advise its removal if you plan to keep the climber controls separate.